feat(sda): move SDA related code to main project and perform clean ups
