+++ src/NebulaAuth/NebulaAuth.csproj
@@ -46,7 +46,6 @@
 	</ItemGroup>
 
 	<ItemGroup>
-		<ProjectReference Include="..\NebulaAuth.LegacyConverter\NebulaAuth.LegacyConverter.csproj" />
 		<ProjectReference Include="..\SteamLibForked\SteamLibForked.csproj" />
 	</ItemGroup>
