+++ src/NebulaAuth/View/Dialogs/SdaPasswordDialog.xaml
@@ -10,7 +10,7 @@
              d:DataContext="{d:DesignInstance other:SdaPasswordDialogVM}"
              Background="{DynamicResource WindowBackground}">
 
-    <Grid MinHeight="200" MinWidth="400">
+    <Grid MinHeight="200" MinWidth="400" MaxWidth="500">
         <Grid.RowDefinitions>
             <RowDefinition Height="Auto" />
             <RowDefinition Height="Auto" />
